backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
cryptography==44.0.2
requests==2.32.3
instagrapi==2.1.3
flask-limiter==3.12
pyjwt==2.10.1
gunicorn==23.0.0
telegram-send==0.37
pytest==8.2.0 